changed import of setup from distutils.core to setuptools
